Move contents_change_handler out of character.h|cpp
